consolidating api's to same controllers
