+++ api/src/index.js
@@ -4,7 +4,6 @@ import "dotenv/config";
 
 import movieRouter from "./routers/movie_router.js";
 import userRouter from "./routers/user_router.js";
-import loginRouter from "./routers/login_router.js";
 
 const app = express();
 const port = process.env.PORT;
@@ -19,7 +18,6 @@ app.get("/", async (req, res) => {
 
 app.use("/movie", movieRouter);
 app.use("/user", userRouter);
-app.use("/user", loginRouter);
 
 app.listen(port, () => {
   console.log(`Server is listening port ${port}`);
