+++ api/src/routers/user_router.js
@@ -1,8 +1,9 @@
 import { Router } from "express";
-import { signUp } from "../controllers/register_controller.js";
+import { signUp, login } from "../controllers/user_controller.js";
 
 const userRouter = Router();
 
 userRouter.post("/signup", signUp);
+userRouter.post("/signin", login);
 
 export default userRouter;
